Refactor cache handling by moving it to a new `vecdb_cache` module and updating VecDb and related components accordingly
