+++ src/vecdb/mod.rs
@@ -5,3 +5,4 @@ pub mod file_filter;
 pub mod structs;
 pub mod vecdb;
 pub mod vecdb_remote;
+mod vecdb_cache;
